+++ .vscode/settings.json
@@ -1,0 +1,26 @@
+{
+	"files.associations": {
+		"minishell_v1.h": "c",
+		"minishell.h": "c",
+		"__node_handle": "c",
+		"__bit_reference": "c",
+		"__functional_base": "c",
+		"algorithm": "c",
+		"bitset": "c",
+		"chrono": "c",
+		"__memory": "c",
+		"functional": "c",
+		"iterator": "c",
+		"limits": "c",
+		"memory": "c",
+		"optional": "c",
+		"ratio": "c",
+		"system_error": "c",
+		"tuple": "c",
+		"type_traits": "c",
+		"vector": "c",
+		"atomic": "c",
+		"locale": "c"
+	},
+	"functionHeaderExtension.headerFilePath": "/Users/ymomen/Desktop/1337-Minishell-42/training/minishell_v1/minishell_v1.h"
+}
