+++ training/minishell_v1/linked_lst.c
@@ -6,69 +6,75 @@
 /*   By: ymomen <ymomen@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/23 17:39:11 by ymomen            #+#    #+#             */
-/*   Updated: 2024/02/29 23:30:37 by ymomen           ###   ########.fr       */
+/*   Updated: 2024/03/02 17:37:24 by ymomen           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell_v1.h"
 
-t_lst *lst_new(char *str)
+t_lst	*lst_new(char *str)
 {
-	t_lst *node;
-	
+	t_lst	*node;
+
 	if (!str || check_arr(str) == 0)
 		return (NULL);
 	node = (t_lst *) malloc(sizeof(t_lst));
 	if (!node)
 		return (NULL);
-	node->value  = str;
+	node->value = str;
 	is_operateur(&node);
 	node->next = NULL;
 	return (node);
 }
-void lst_add_back(t_lst **head, t_lst *node)
+
+void	lst_add_back(t_lst **head, t_lst *node)
 {
-	t_lst * tmp;
+	t_lst	*tmp;
+
 	if (!head || !node)
 		return ;
 	if (!*head)
 	{
 		*head = node;
-		return;
+		(*head)->next = NULL;
+		return ;
 	}
 	else
 	{
 		tmp = *head;
-		while(tmp->next)
+		while (tmp->next)
 			tmp = tmp->next;
 		tmp->next = node;
 		node->next = NULL;
-	return ;	
+		return ;
 	}
 }
-int size_lst(t_lst **head)
+
+int	size_lst(t_lst **head)
 {
-	t_lst *tmp;
-	int i;
+	t_lst	*tmp;
+	int		i;
 
 	if (!head || !*head)
 		return (0);
 	i = 0;
 	tmp = *head;
-	while(tmp)
+	while (tmp)
 	{
 		i++;
 		tmp = tmp->next;
 	}
 	return (i);
 }
-void lst_clear(t_lst *head)
+
+void	lst_clear(t_lst *head)
 {
-	t_lst *node;
+	t_lst	*node;
+
 	if (!head)
 		return ;
 	node = head;
-	while(head)
+	while (head)
 	{
 		head = head ->next;
 		free(node->value);
@@ -76,45 +82,31 @@ void lst_clear(t_lst *head)
 		node = head;
 	}
 }
-void *pop_last(t_lst **stack)
+
+void	*pop_last(t_lst **stack)
 {
-	void  *ptr = NULL;
-	t_lst *tmp;
-	t_lst *prv;
+	void	*ptr;
+	t_lst	*tmp;
+	t_lst	*prv;
+
 	if (!stack || !*stack)
 		return (NULL);
-	
 	tmp = *stack;
 	if (!(*stack)->next)
 	{
 		ptr = (*stack)->value;
 		free(*stack);
-		*stack= NULL; 
-		return(ptr);
+		return (ptr);
 	}
-	while((*stack)->next)
+	while ((*stack)->next)
 	{
 		prv = *stack;
 		*stack = (*stack)->next;
 	}
 	ptr = (*stack)->value;
-		free(*stack);
-		*stack= NULL;
-		prv->next = NULL;
+	free(*stack);
+	*stack = NULL;
+	prv->next = NULL;
 	*stack = tmp;
 	return (ptr);
-}
-
-void	ft_lstadd_front(t_lst **lst, t_lst *new)
-{
-	new->next = *lst;
-	*lst = new;
-}
-t_lst *lastone(t_lst *head)
-{
-	if (!head)
-		return (NULL);
-	while (head->next)
-		head = head->next;
-	return(head);
 }
